Refactor metal Kernels
